Add files via upload
Suicide in Nigeria has become an enigma that is slowly eating into the lives of the citizens. In Nigeria, according to the World Health Organization, WHO, an estimated 800,000 people die by suicide every year. Nigeria is now ranked the 10th in Africa and 67th in the world with a crude suicide rate of 9.5 per 100,000 population. Due to the rate at which the number of suicides is growing yields the motivation for carrying out this project. This research is carried out to predict the causes of suicide in Nigeria by using machine learning algorithms. By identifying the root causes behind suicide, we can create awareness to prevent suicide cases and rectify the predicted causes of suicide. In this research, we have analyzed the pattern of suicide cases and predict the causes of future suicides by using Bag of Words technique and Logistic regression for the first part of the system and XGBoost and Random Forest machine learning algorithms for the second part of the system to predict these causes.